+++ src/opender/auxiliary_funcs/flipflop.py
