+++ src/opender/auxiliary_funcs/low_pass_filter.py
@@ -12,7 +12,7 @@
 #   prior written permission.
 
 
-from . import der
+from opender import der
 
 
 class LowPassFilter:
