+++ src/opender/auxiliary_funcs/time_delay.py
@@ -12,7 +12,7 @@
 #   prior written permission.
 
 
-from . import der
+from opender import der
 
 
 class TimeDelay:
