+++ src/opender/enter_service_perf.py
@@ -11,13 +11,7 @@
 #   to endorse or promote products derived from this software without specific
 #   prior written permission.
 
-from .time_delay import TimeDelay
-from .ramping import Ramping
-from . import der
-from .flipflop import FlipFlop
-from .cond_delay import ConditionalDelay
-from . import operating_condition_input_processing
-import numpy as np
+from opender.auxiliary_funcs.ramping import Ramping
 
 
 # %%
