+++ src/opender/reactive_power_support_funcs/volt_var.py
@@ -12,7 +12,7 @@
 #   prior written permission.
 
 
-from .low_pass_filter import LowPassFilter
+from opender.auxiliary_funcs.low_pass_filter import LowPassFilter
 
 
 class VoltVAR:
