+++ src/opender/reactive_power_support_funcs/watt_var.py
@@ -11,7 +11,7 @@
 #   to endorse or promote products derived from this software without specific
 #   prior written permission.
 
-from .low_pass_filter import LowPassFilter
+from opender.auxiliary_funcs.low_pass_filter import LowPassFilter
 
 
 class WattVAR:
